Refactor app into a package
