+++ app/__main__.py
@@ -1,0 +1,5 @@
+from app.main import run_app
+
+
+if __name__ == '__main__':
+    run_app()
