+++ pyproject.toml
@@ -13,6 +13,7 @@ aioauth-client = "^0.16.2"
 aiohttp-session = "^2.7"
 cryptography = "^2.7"
 aiohttp = "^3.8.3"
+django-environ = "^0.9.0"
 
 [tool.poetry.group.dev.dependencies]
